pyproject.toml
[tool.poetry]
name = "noise-image-generator"
version = "0.1.0"
description = ""
authors = ["Malte Eckhoff <malte.eckhoff@lusini.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
